+++ rest_framework/fields.py
@@ -535,7 +535,10 @@ class HyperlinkedRelatedField(RelatedField):
             }
         request = self.context.get('request', None)
         format = self.format or self.context.get('format', None)
-        kwargs = {self.pk_url_kwarg: obj.pk}
+        pk = getattr(obj, 'pk', None)
+        if pk is None:
+            return
+        kwargs = {self.pk_url_kwarg: pk}
         try:
             return reverse(view_name, kwargs=kwargs, request=request, format=format)
         except: