+++ docs/topics/release-notes.md
@@ -4,6 +4,17 @@
 >
 > &mdash; Eric S. Raymond, [The Cathedral and the Bazaar][cite].
 
+## Master
+
+* Support for `read_only_fields` on `ModelSerializer` classes.
+
+## 2.1.2
+
+**Date**: 9th Nov 2012
+
+* **Filtering support.**
+* Bugfix: Support creation of objects with reverse M2M relations.
+
 ## 2.1.1
 
 **Date**: 7th Nov 2012